requirements.txt
	fastapi==0.103.0
	uvicorn==0.23.2
	PyPDF2==3.0.1
	Pillow==11.2.1
	pytesseract==0.3.10
	scispacy==0.5.3
	spacy==3.6.1
	fuzzywuzzy==0.18.0
	Levenshtein==0.21.1
	transformers==4.35.0
	torch==2.0.1
	aiohttp==3.8.5
	rasa==3.6.0

